Fix url navigations and reorganise folder
