+++ django_material_demo/cms/polls/views.py
@@ -1,5 +1,4 @@
 from django import forms
-from django.contrib.admin.views.decorators import staff_member_required
 from django.contrib.auth.forms import UserChangeForm
 from django.core.exceptions import NON_FIELD_ERRORS, ValidationError
 from django.db.models import Count
@@ -7,10 +6,7 @@ from django.forms import (BaseInlineFormSet, BooleanField, EmailField,
                           ModelForm, model_to_dict)
 from django.forms.widgets import CheckboxInput, RadioSelect
 from django.http import HttpResponse
-from django.shortcuts import redirect
-from django.utils.decorators import method_decorator
 from django.views import generic
-from django.views.generic.edit import FormView
 from django_filters import CharFilter, NumberFilter
 from library.django_superform import (ForeignKeyFormField, InlineFormSetField,
                                       SuperModelForm)
@@ -22,10 +18,9 @@ from material.frontend.views import (CreateModelView, DetailModelView,
 from polls.models import (Attachment, Choice, File, Question, QuestionFollower,
                           User, Vote)
 
-from .models import Settings
-from .utils import (FieldDataMixin, FormSetForm, GetParamAsFormDataMixin,
-                    ListFilterView, NestedModelFormField, SearchAndFilterSet,
-                    get_html_list)
+from ..utils import (FieldDataMixin, FormSetForm, GetParamAsFormDataMixin,
+                     ListFilterView, NestedModelFormField, SearchAndFilterSet,
+                     get_html_list)
 
 
 class FileViewSet(ModelViewSet):
@@ -464,34 +459,3 @@ class PasswordChangeView(generic.View):
 class PasswordChangeDoneView(generic.View):
     def get(self, request, *args, **kwargs):
         return HttpResponse('Password changed!')
-
-
-class SettingsForm(forms.Form):
-    primary_color = forms.CharField(label='Primary color', required=False)
-    primary_color_light = forms.CharField(
-        label='Primary color light', required=False)
-    primary_color_dark = forms.CharField(
-        label='Primary color dark', required=False)
-    secondary_color = forms.CharField(label='Secondary color', required=False)
-    secondary_color_light = forms.CharField(
-        label='Secondary color light', required=False)
-    success_color = forms.CharField(label='Success color', required=False)
-    error_color = forms.CharField(label='Error color', required=False)
-    link_color = forms.CharField(label='Link color', required=False)
-
-
-@method_decorator(staff_member_required(login_url='login'), name='dispatch')
-class SettingsView(FormView):
-    form_class = SettingsForm
-    template_name = 'cms/settings.html'
-
-    def get_initial(self):
-        return model_to_dict(Settings(session=self.request.session))
-
-    def form_valid(self, form):
-        settings = Settings(session=self.request.session)
-        for k, v in form.cleaned_data.items():
-            setattr(settings, k, v)
-
-        settings.save()
-        return redirect('settings')
