+++ django_material_demo/cms/settings/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+from material.frontend.apps import ModuleMixin
+
+
+class CmsSettingsConfig(ModuleMixin, AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    icon = '<i class="material-icons">settings</i>'
+    name = 'cms.settings'
+    label = 'settings'
