+++ django_material_demo/cms/settings/templates/cms/settings.html
@@ -9,7 +9,6 @@
 {% endblock %}
 
 {% block content %}
-{% block left-panel %}
 <div class="left-panel">
     <div class="card">
         <form method="POST" enctype="multipart/form-data">
@@ -136,6 +135,4 @@
         </form>
     </div>
 </div>
-{% endblock left-panel %}
-{% block right-panel %}{% endblock right-panel %}
 {% endblock %}
