+++ django_material_demo/cms/settings/templatetags/settings.py
