+++ django_material_demo/cms/settings/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from .views import SettingsView
+
+app_name = 'settings'
+urlpatterns = [
+    path('', SettingsView.as_view(), name='index'),
+]
