+++ django_material_demo/cms/urls.py
@@ -1,0 +1,6 @@
+from django.urls import include, path
+from material.frontend import urls as frontend_urls
+
+urlpatterns = [
+    path('', include(frontend_urls)),
+]
