+++ django_material_demo/django_material_demo/settings.py
@@ -41,6 +41,7 @@ if CSRF_TRUSTED_ORIGINS_STR:
 
 INSTALLED_APPS = [
     'cms.polls.apps.CmsPollsConfig',
+    'cms.settings.apps.CmsSettingsConfig',
     'polls.apps.PollsConfig',
     'material',
     'material.frontend',
