+++ django_material_demo/polls/apps.py
@@ -1,5 +1,4 @@
 from django.apps import AppConfig
-from material.frontend.apps import ModuleMixin
 
 
 class PollsConfig(AppConfig):
