+++ django_material_demo/polls/templates/polls/home.html
@@ -9,7 +9,7 @@
     {% if latest_question_list %}
         <ul>
             {% for question in latest_question_list %}
-            <li><a href="{% url 'polls:question' question.id %}">{{ question.question_text }}</a></li>
+            <li><a href="{% url 'app_polls:question' question.id %}">{{ question.question_text }}</a></li>
             {% endfor %}
         </ul>
     {% else %}
