+++ django_material_demo/polls/templates/polls/includes/top_nav.html
@@ -1,1 +1,1 @@
-<a href="/admin">Admin portal</a> | <a href="/polls/admin">Admin portal (django-material) </a>
+<a href="/admin">Admin portal</a> | <a href="/cms/polls">Admin portal (django-material) </a>
