+++ django_material_demo/polls/tests.py
@@ -54,7 +54,7 @@ class QuestionIndexViewTests(TestCase):
         """
         If no questions exist, an appropriate message is displayed.
         """
-        response = self.client.get(reverse('polls:index'))
+        response = self.client.get(reverse('app_polls:index'))
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "No polls are available.")
         self.assertQuerysetEqual(response.context['latest_question_list'], [])
@@ -65,7 +65,7 @@ class QuestionIndexViewTests(TestCase):
         index page.
         """
         question = create_question(question_text="Past question.", days=-30)
-        response = self.client.get(reverse('polls:index'))
+        response = self.client.get(reverse('app_polls:index'))
         self.assertQuerysetEqual(
             response.context['latest_question_list'],
             [question],
@@ -77,7 +77,7 @@ class QuestionIndexViewTests(TestCase):
         the index page.
         """
         create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
+        response = self.client.get(reverse('app_polls:index'))
         self.assertContains(response, "No polls are available.")
         self.assertQuerysetEqual(response.context['latest_question_list'], [])
 
@@ -88,7 +88,7 @@ class QuestionIndexViewTests(TestCase):
         """
         question = create_question(question_text="Past question.", days=-30)
         create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
+        response = self.client.get(reverse('app_polls:index'))
         self.assertQuerysetEqual(
             response.context['latest_question_list'],
             [question],
@@ -100,7 +100,7 @@ class QuestionIndexViewTests(TestCase):
         """
         question1 = create_question(question_text="Past question 1.", days=-30)
         question2 = create_question(question_text="Past question 2.", days=-5)
-        response = self.client.get(reverse('polls:index'))
+        response = self.client.get(reverse('app_polls:index'))
         self.assertQuerysetEqual(
             response.context['latest_question_list'],
             [question2, question1],
@@ -115,7 +115,7 @@ class QuestionDetailViewTests(TestCase):
         """
         future_question = create_question(
             question_text='Future question.', days=5)
-        url = reverse('polls:detail', args=(future_question.id,))
+        url = reverse('app_polls:detail', args=(future_question.id,))
         response = self.client.get(url)
         self.assertEqual(response.status_code, 404)
 
@@ -126,6 +126,6 @@ class QuestionDetailViewTests(TestCase):
         """
         past_question = create_question(
             question_text='Past Question.', days=-5)
-        url = reverse('polls:detail', args=(past_question.id,))
+        url = reverse('app_polls:detail', args=(past_question.id,))
         response = self.client.get(url)
         self.assertContains(response, past_question.question_text)
